Fix bug when calculating next run, result shows 1 minute ahead of expected. Add support for trailing whitespaces and tab separation
Bug - if cron input is set to to run every min (* *) and the current time is HH:15:59 (important part is 59 seconds) then the next run will show as HH:16 as there is no seconds within the current minute to run the job (and the cron library supports definition to a level of seconds).
This is correct cron behaviour but seeing as the problem spec did not mention seconds, there is reasonable expectation that they would not effect the output.